refactor: split GameScene.ts and TalentTreeOverlay.ts into smaller files
- Extract server entity sync logic from GameScene → ServerEntitySync.ts (281 lines)
- Extract skill slot panel from TalentTreeOverlay → SkillSlotPanel.ts (358 lines)
- GameScene: 781 → 537 lines
- TalentTreeOverlay: 737 → 433 lines
- All files now within 200-400 line target range

Closes #183
